Move linter module to reducers
